+++ scglobals/static/tx/src/app/services/create-couch-views.js
@@ -30,6 +30,19 @@ export default class CreateCouchViews {
             }
         }
       },
+      // for inventory
+      {
+        name: 'inventory',
+        mapFun: function(doc){
+            if (doc.doc_type && doc.doc_type === "shipment"){
+              // see all-transactions comment
+              doc.transactions.forEach(function(transaction,i){
+                emit([doc.from_location_name,transaction.item_name,transaction.item_category_name,transaction.item_lot_expiration,transaction.item_lot_num], -1*transaction.qty);
+                emit([doc.to_location_name,transaction.item_name,transaction.item_category_name,transaction.item_lot_expiration,transaction.item_lot_num], transaction.qty);
+              });
+            }
+        }
+      },
       // for item search http://stackoverflow.com/questions/5456682/return-unique-values-by-key-in-couchdb
       {
         name: 'items-by-name',